fixed vad service + added logic of angle calculation

+++ src/dto/request/audio.rs
@@ -2,6 +2,6 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
 pub struct AudioData {
-    pub mic1: Vec<u8>,
-    pub mic2: Vec<u8>,
+    pub mic1: Vec<i16>,
+    pub mic2: Vec<i16>,
 }
